// pc/src/services/apis/models/ExtZsProjectOperationVo.ts
export interface ExtZsProjectOperationVo {
    /**
     * 项目名称
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    name?: string;
    /**
     * 招引单位
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    zoneName?: string;
    /**
     * 投资方名称
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    investor?: string;
    /**
     * 项目选址位置
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    projectAddress?: string;
    /**
     * 签约日期
     * @type {Date}
     * @memberof ExtZsProjectOperationVo
     */
    signedStatDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    bIndustry?: string;
    /**
     * 是否科创项目
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    isKcProj?: string;
    /**
     * 科创项目认定条件
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    kcProjTj?: string;
    /**
     * QFLP外资项目
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    isQflp?: string;
    /**
     * 
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    uCode?: string;
    /**
     * 注册日期
     * @type {Date}
     * @memberof ExtZsProjectOperationVo
     */
    regDate?: Date;
    /**
     * 主要产品、产能及主要建设内容
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    desc?: string;
    /**
     * 行业分类代码
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    industryCode?: string;
    /**
     * 行业分类
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    industryName?: string;
    /**
     * 产业方向
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    projType?: string;
    /**
     * 项目类型, t_proj_type
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    projTypeLabel?: string;
    /**
     * 总投资
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    investMoney?: string;
    /**
     * 固定资产投资（万元）
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    fixedInvest?: string;
    /**
     * 批准部门及文号
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    pzwh?: string;
    /**
     * 批准日期
     * @type {Date}
     * @memberof ExtZsProjectOperationVo
     */
    pzrq?: Date;
    /**
     * 成效情况
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    cxqk?: string;
    /**
     * 备案核准日期
     * @type {Date}
     * @memberof ExtZsProjectOperationVo
     */
    checkStatDate?: Date;
    /**
     * 开工作证材料
     * @type {Array<string>}
     * @memberof ExtZsProjectOperationVo
     */
    kgzzcl?: Array<string>;
    /**
     * 竣工佐证材料
     * @type {Array<string>}
     * @memberof ExtZsProjectOperationVo
     */
    jgzzcl?: Array<string>;
    /**
     * 质态评估评审结果
     * @type {Array<string>}
     * @memberof ExtZsProjectOperationVo
     */
    ztpgzzcl?: Array<string>;
    /**
     * 开工日期
     * @type {Date}
     * @memberof ExtZsProjectOperationVo
     */
    startDate?: Date;
    /**
     * 竣工
     * @type {Date}
     * @memberof ExtZsProjectOperationVo
     */
    endDate?: Date;
    /**
     * 招引单位
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    department?: string;
    /**
     * 
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    pType?: string;
    /**
     * 是否为签约工业、服务业项目转科创项目
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    kgIsZkc?: string;
    /**
     * 符合科创证明材料
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    kczzcl?: string;
    /**
     * 人才佐证材料
     * @type {Array<string>}
     * @memberof ExtZsProjectOperationVo
     */
    rczzcl?: Array<string>;
    /**
     *  企业人员总数
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    qyTotalNum?: string;
    /**
     * 企业社保2个月以上人数
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    qySbNum?: string;
    /**
     * 企业在泰研发人数
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    qyTzyfNum?: string;
    /**
     * 企业当年度研发投入（万元）
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    qyYfMoney?: string;
    /**
     * 参保作证材料/专利证书或受理通知书证明材料
     * @type {Array<string>}
     * @memberof ExtZsProjectOperationVo
     */
    cbzzcl?: Array<string>;
    /**
     * 1-服务业 2-制造业
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    bindustry?: string;
    /**
     * 内外资
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    ptype?: string;
    /**
     * 统一社会信用代码
     * @type {string}
     * @memberof ExtZsProjectOperationVo
     */
    ucode?: string;
}